Test IDM-to-IDM trust with ipalab-config
This patch adds instructions, configuration files, and playbooks to
prepare an environment and create an IDM-to-IDM trust using Ansible.

The new 'establish-trust.yaml' provided tries to use Ansible collections
to the greater extense possible, instead of 'ansible.bulitin.shell' and
makes use of ansible-freeipa to all IPA commands currently supported.

Note that the use of ipalab-config is self-contained, not using any
other file from the repository.